JavaHW 02 added test and modified operands to double

+++ JavaHW_02/src/javahw_02/Evaluator.java
@@ -16,7 +16,7 @@ class Evaluator {
         this.operator = operator;
     }
 
-    int evaluate(Operands o) {
+    double evaluate(Operands o) {
         if (operator.equals("+")){
             return o.getFirstOp() + o.getSecondOp();
         }
